refactor: move auth tools to separate location at root
