+++ web/frontend/src/utils/auth.ts
@@ -1,4 +1,4 @@
-import { ID } from './../../../../types/configuration';
+import { ID } from './../types/configuration';
 
 export function isManager(formID: ID, authorization: Map<String, String[]>, isLogged: boolean) {
   return (
@@ -7,5 +7,13 @@ export function isManager(formID: ID, authorization: Map<String, String[]>, isLo
     authorization.get('election').includes('create') && // must be able to create elections
     authorization.has(formID) &&
     authorization.get(formID).includes('own') // must own the election
+  );
+}
+
+export function isVoter(formID: ID, authorization: Map<String, String[]>, isLogged: boolean) {
+  return (
+    isLogged && // must be logged in
+    authorization.has(formID) &&
+    authorization.get(formID).includes('vote') // must be able to vote in the election
   );
 }
